search-service/pyproject.toml
[project]
name = "search-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beanie>=1.30.0",
    "elasticsearch[async]>=9.0.2",
    "fastapi[standard]>=0.115.14",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "yarl>=1.20.1",
]
